updated: convention plugins directories
