+++ build-logic/convention/src/main/kotlin/dev/enesky/build_logic/convention/plugins/common/FeaturePlugin.kt
@@ -1,4 +1,4 @@
-package dev.enesky.build_logic.plugins.common
+package dev.enesky.build_logic.convention.plugins.common
 
 import dev.enesky.build_logic.convention.libs
 import org.gradle.api.Plugin
