+++ build-logic/convention/src/main/kotlin/dev/enesky/build_logic/convention/plugins/common/SigningConfigPlugin.kt
@@ -1,4 +1,4 @@
-package dev.enesky.build_logic.plugins.common
+package dev.enesky.build_logic.convention.plugins.common
 import com.android.build.gradle.internal.dsl.BaseAppModuleExtension
 import dev.enesky.build_logic.convention.createSigningConfigFromProperties
 import org.gradle.api.Plugin
