+++ build-logic/convention/src/main/kotlin/dev/enesky/build_logic/convention/plugins/common/TestPlugin.kt
@@ -1,4 +1,4 @@
-package dev.enesky.build_logic.plugins.common
+package dev.enesky.build_logic.convention.plugins.common
 
 import com.android.build.gradle.TestExtension
 import dev.enesky.build_logic.convention.configureKotlinAndroid
