--- build-logic/convention/src/main/kotlin/dev/enesky/build_logic/plugins/app/AppJacocoPlugin.kt
@@ -1,21 +1,0 @@
-package dev.enesky.build_logic.plugins.app
-
-import com.android.build.api.variant.ApplicationAndroidComponentsExtension
-import dev.enesky.build_logic.convention.configureJacoco
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-import org.gradle.kotlin.dsl.getByType
-
-class AppJacocoPlugin : Plugin<Project> {
-    override fun apply(target: Project) {
-        with(target) {
-            with(pluginManager) {
-                apply("org.gradle.jacoco")
-                apply("com.android.application")
-            }
-            val extension = extensions.getByType<ApplicationAndroidComponentsExtension>()
-            configureJacoco(extension)
-        }
-    }
-
-}
